+++ CHANGELOG.md
@@ -14,6 +14,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 - [shelly]: The Input component is not added when is not enabled in the device web page.
 - [shelly]: Input component configured like Button are now supported as momentary switch if exposeInputEvent=momentary.
+- [shelly]: Configure Light and Rgb components.
 
 ### Verified